Move `HttpClientStub` to its own file
